Fix todo-list-component test
This commit puts the code for the todo-list-component in the right file. Previously, it was in todo.service.spec.ts when it should have been in todo-list-component.spec.ts
