--- client/src/app/todo/todo.service.spec.ts
@@ -1,132 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { FormsModule } from '@angular/forms';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatOptionModule } from '@angular/material/core';
-import { MatDividerModule } from '@angular/material/divider';
-import { MatExpansionModule } from '@angular/material/expansion';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
-import { MatListModule } from '@angular/material/list';
-import { MatRadioModule } from '@angular/material/radio';
-import { MatSelectModule } from '@angular/material/select';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { MatTooltipModule } from '@angular/material/tooltip';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { RouterTestingModule } from '@angular/router/testing';
-import { Observable } from 'rxjs';
-import { MockTodoService } from '../../testing/todo.service.mock';
-import { Todo } from './todo';
-import { TodoListComponent } from './todo-list.component';
-import { TodoService } from '././todo.service';
-
-const COMMON_IMPORTS: any[] = [
-  FormsModule,
-  MatCardModule,
-  MatFormFieldModule,
-  MatSelectModule,
-  MatOptionModule,
-  MatButtonModule,
-  MatInputModule,
-  MatExpansionModule,
-  MatTooltipModule,
-  MatListModule,
-  MatDividerModule,
-  MatRadioModule,
-  MatSnackBarModule,
-  BrowserAnimationsModule,
-  RouterTestingModule,
-];
-
-describe('TodoListComponent', () => {
-
-  let todoList: TodoListComponent;
-  let fixture: ComponentFixture<TodoListComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [COMMON_IMPORTS],
-      declarations: [TodoListComponent],
-      // providers:    [ UserService ]  // NO! Don't provide the real service!
-      // Provide a test-double instead
-      // This MockerUserService is defined in client/testing/user.service.mock.
-      providers: [{ provide: TodoService, useValue: new MockTodoService() }]
-    });
-  });
-
-  beforeEach(waitForAsync(() => {
-    TestBed.compileComponents().then(() => {
-      // Create a "fixture" of the UserListComponent. that
-      // allows us to get an instance of the component
-      // (userList, below) that we can "control" in
-      // the tests.
-      fixture = TestBed.createComponent(TodoListComponent);
-      todoList = fixture.componentInstance;
-      fixture.detectChanges();
-    });
-  }));
-
-  it('contains all the todos', () => {
-    expect(todoList.serverFilteredTodos.length).toBe(3);
-  });
-
-  it('contains all the todos that are completed', () => {
-    expect(todoList.serverFilteredTodos.some((todo: Todo) => todo.status === true)).toBe(false);
-  });
-});
-
-/*
- * This test is a little odd, but illustrates how we can use "stubbing"
- * to create mock objects (a service in this case) that be used for
- * testing. Here we set up the mock UserService (userServiceStub) so that
- * _always_ fails (throws an exception) when you request a set of users.
- *
- * So this doesn't really test anything meaningful in the context of our
- * code (I certainly wouldn't copy it), but it does illustrate some nice
- * testing tools. Hopefully it's useful as an example in that regard.
- */
-describe('Misbehaving Todo List', () => {
-  let todoList: TodoListComponent;
-  let fixture: ComponentFixture<TodoListComponent>;
-
-  let todoServiceStub: {
-    getTodos: () => Observable<Todo[]>;
-    getTodosFiltered: () => Observable<Todo[]>;
-  };
-
-  beforeEach(() => {
-    // stub UserService for test purposes
-    todoServiceStub = {
-      getTodos: () => new Observable(observer => {
-        observer.error('Error-prone observable');
-      }),
-      getTodosFiltered: () => new Observable(observer => {
-        observer.error('Error-prone observable');
-      })
-    };
-
-    TestBed.configureTestingModule({
-      imports: [COMMON_IMPORTS],
-      declarations: [TodoListComponent],
-      // providers:    [ UserService ]  // NO! Don't provide the real service!
-      // Provide a test-double instead
-      providers: [{ provide: TodoService, useValue: todoServiceStub }]
-    });
-  });
-
-  beforeEach(waitForAsync(() => {
-    TestBed.compileComponents().then(() => {
-      fixture = TestBed.createComponent(TodoListComponent);
-      todoList = fixture.componentInstance;
-      fixture.detectChanges();
-    });
-  }));
-
-  it('fails to load todos if we do not set up a TodoListService', () => {
-    // Since calling both getUsers() and getUsersFiltered() return
-    // Observables that then throw exceptions, we don't expect the component
-    // to be able to get a list of users, and serverFilteredUsers should
-    // be undefined.
-    expect(todoList.serverFilteredTodos).toBeUndefined();
-  });
-});
